LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-remove-head-expected.txt


FAIL Meta color-scheme applies. assert_equals: expected (string) "dark" but got (undefined) undefined
FAIL Initial value after removing head including meta color-scheme. assert_equals: expected (string) "normal" but got (undefined) undefined
